pyproject.toml
[project]
name = "autothreat-ai"
version = "0.1.0"
description = "An automated threat modeling platform built with Google Agent Development Kit (ADK) and Gemini."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk[a2a]>=1.22.1",
    "google-genai>=1.59.0",
    "markdown-pdf>=1.10",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
]
